seed/requirements.txt
pocketbase
python-dotenv
pandas